+++ pykit/mark.py
@@ -8,12 +8,10 @@ from typing import Any, Awaitable, Callable
 
 from pykit.check import check
 from pykit.err import InpErr
-from pykit.fcode import code
 from pykit.query import UpdQuery
 from pykit.types import T
 
 
-@code("mark-err")
 class MarkErr(Exception):
     pass