+++ pykit/proc.py
@@ -6,7 +6,7 @@ from xml.dom import NotFoundErr
 
 from result import Err, Ok
 
-from pykit.err import ValueErr
+from pykit.err import ValErr
 from pykit.log import log
 from pykit.res import Res
 
@@ -59,7 +59,7 @@ class ProcGroup:
         interfacing uses.
         """
         if not self._can_register_by_limit():
-            return Err(ValueErr(
+            return Err(ValErr(
                 "cannot register a new process:"
                 f" limit {self._max_procs} is exceeded"))
 
@@ -75,13 +75,13 @@ class ProcGroup:
                 f"proc {proc} has been started, but the pid is unassigned"))
         if self.has(proc.pid):
             proc.kill()
-            return Err(ValueErr(
+            return Err(ValErr(
                 "new process is started with the same pid as registered"
                 " one => kill new process"))
 
         if key:
             if key in self._key_to_pid:
-                return Err(ValueErr(f"key {key} is already registered"))
+                return Err(ValErr(f"key {key} is already registered"))
             self._key_to_pid[key] = proc.pid
 
         self._procs[proc.pid] = (proc, parent_pipe)
@@ -187,5 +187,5 @@ class ProcGroup:
         proc, pipe = self._procs[pid]
         if not proc.is_alive():
             self.try_deregister(pid).unwrap()
-            return Err(ValueErr("process is closed"))
+            return Err(ValErr("process is closed"))
         return Ok((proc, pipe))